Created EventCreator, EventModifier and SingleEventAccessor interfaces as well as Hilt bindings for each.
