+++ trackr-app/app/src/main/java/com/trackr/trackr_app/manager/EventCreator.kt
@@ -1,17 +1,20 @@
 package com.trackr.trackr_app.manager
 
+import com.trackr.trackr_app.model.TrackrEvent
+import com.trackr.trackr_app.viewmodels.AddScreenViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityComponent
-import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.components.SingletonComponent
+import java.time.LocalDate
 
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class EventCreator {
-    @Binds
-    abstract fun bindEventCreator(
-        eventCreatorImpl: EventManager
-    ): EventCreatorInterface
+/*
+An interface depended on by any view model that needs to add new events to the database.
+Any Manager class that creates and adds events to the database should implement this interface.
+ */
+interface EventCreator {
+    suspend fun addEvent(firstName: String, lastName: String, eventType: Int, chosenReminder: String,
+                 eventDate: LocalDate)
 }
+
