+++ trackr-app/app/src/main/java/com/trackr/trackr_app/manager/EventCreatorBinding.kt
@@ -1,0 +1,15 @@
+package com.trackr.trackr_app.manager
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class EventCreatorBinding {
+    @Binds
+    abstract fun bindEventCreator(
+        eventCreatorImpl: EventManager
+    ): EventCreator
+}
