+++ trackr-app/app/src/main/java/com/trackr/trackr_app/manager/EventManager.kt
@@ -21,7 +21,7 @@ class EventManager @Inject constructor(
     private val eventRepository: EventRepository,
     private val eventNotificationManager: EventNotificationManager,
     private val personManager: PersonManager,
-) : EventCreatorInterface, EventModifierInterface, SingleEventAccessorInterface {
+) : EventCreator, EventModifier, SingleEventAccessor {
     private fun createEvent(id: String, type: Int, date: Long, firstYear: Int, reminderInterval: Int,
                             reminderStrategy: Int): TrackrEvent {
         return TrackrEvent(id, type, date, firstYear, reminderInterval, reminderStrategy)
