+++ trackr-app/app/src/main/java/com/trackr/trackr_app/manager/EventModifier.kt
@@ -6,7 +6,7 @@ import java.time.LocalDate
 An interface depended on by any view model that needs to modify existing events in the database.
 Any Manager class that can edit or delete events in the database should implement on this interface.
  */
-interface EventModifierInterface {
+interface EventModifier {
     suspend fun editEvent(eventID: String, reminderInt: Int, eventDate: LocalDate, eventType: Int,
                   personName: String, eventName: String)
     suspend fun deleteEvent(eventID: String)
