+++ trackr-app/app/src/main/java/com/trackr/trackr_app/viewmodels/AddScreenViewModel.kt
@@ -4,9 +4,7 @@ import android.media.metrics.Event
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.*
-import com.trackr.trackr_app.manager.EventManager
 import com.trackr.trackr_app.manager.EventCreator
-import com.trackr.trackr_app.manager.EventCreatorInterface
 import com.trackr.trackr_app.manager.PersonManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -15,13 +13,13 @@ import javax.inject.Inject
 
 /**
  * The view model for the AddScreen which manages the data that appears on the AddScreen page
- * @param eventManager an instance of the EventManager class that is used to create new events
- * and add them to the database
+ * @param eventCreator an instance of a class which implements EventCreator that is used to create
+ * new events and add them to the database
  */
 @HiltViewModel
 class AddScreenViewModel @Inject constructor(
         private val personManager: PersonManager,
-        private val eventManager: EventCreatorInterface,
+        private val eventCreator: EventCreator,
         state: SavedStateHandle
 ): ViewModel() {
 
@@ -132,7 +130,7 @@ class AddScreenViewModel @Inject constructor(
      */
     fun addEvent() = viewModelScope.launch {
         //Add the current user to the database
-        eventManager.addEvent(firstName.value, lastName.value,
+        eventCreator.addEvent(firstName.value, lastName.value,
                 eventType, chosenReminder.value, eventDate.value)
     }
 }
